College Project
So for this project i was given an original piece of code i had to debug and fix (Task4a_currency_conversion.py i was told to upgrade this code so that it works properly and has security requirements. The next part of the project was to use the data provide to show trends and patterns in the data i couldn't manage how to configure this with the dates.

Note the original pieces provided to me were: Task4a_data.csv , Task4a_currency_conversion.py